// prisma/interfaces.ts
// This file was auto-generated by prisma-generator-typescript-interfaces

export type user_role = "ADMIN" | "DOSEN" | "PENERBIT";

export interface Major {
  id: number;
  major_name: string;
  major_description: string | null;
  createdAt: Date;
  updatedAt: Date;
  deleted: boolean;
  users?: User[];
}

export interface User {
  id: number;
  name: string;
  email: string;
  password: string;
  role: user_role;
  major_id: number | null;
  major?: Major | null;
  address: string | null;
  phone_number: bigint | null;
  profile_pic: string | null;
  publications?: Publication[];
  publisherPublication?: Publication[];
  activities?: PublicationActivity[];
  createdAt: Date;
  updatedAt: Date;
  deleted: boolean;
}

export interface PublicationStatus {
  id: number;
  status_name: string;
  status_description: string | null;
  publications?: Publication[];
  activities?: PublicationActivity[];
  createdAt: Date;
  updatedAt: Date;
  deleted: boolean;
}

export interface Publication {
  id: number;
  user_id: number;
  publisher_id: number | null;
  lecturer?: User;
  publisher?: User | null;
  status?: PublicationStatus;
  activities?: PublicationActivity[];
  publication_ticket: string;
  publication_title: string;
  publication_document: string;
  current_status_id: number;
  supporting_url: string | null;
  createdAt: Date;
  updatedAt: Date;
  deleted: boolean;
}

export interface PublicationActivity {
  id: number;
  publication_id: number;
  user_id: number;
  publication_status_id: number;
  publication?: Publication;
  user?: User;
  status?: PublicationStatus;
  publication_notes: string | null;
  supporting_url: string | null;
  createdAt: Date;
  updatedAt: Date;
  deleted: boolean;
}